feat(database): add database service & connection
