+++ articlehub-api/internal/database/database.go
@@ -1,15 +1,9 @@
 package database
 
 import (
-	"context"
 	"database/sql"
-	"fmt"
-	"log"
-	"os"
-	"strconv"
-	"time"
 
-	"articlehub-api/internal/models"
+	"articlehub-api/internal/database/repository"
 
 	_ "github.com/jackc/pgx/v5/stdlib"
 	_ "github.com/joho/godotenv/autoload"
@@ -17,6 +11,8 @@ import (
 
 // Service represents a service that interacts with a database.
 type Service interface {
+	UserRepo() repository.UserRepository
+
 	// Health returns a map of health status information.
 	// The keys and values in the map are service-specific.
 	Health() map[string]string
@@ -24,206 +20,43 @@ type Service interface {
 	// Close terminates the database connection.
 	// It returns an error if the connection cannot be closed.
 	Close() error
-
-	// CRUD Operations
-	CreateUser(ctx context.Context, user *models.User) error
-	GetUsers(ctx context.Context) ([]models.User, error)
-	GetUserById(ctx context.Context, id int) (*models.User, error)
-	UpdateUser(ctx context.Context, id int, user *models.User) error
-	DeleteUser(ctx context.Context, id int) error
 }
 
 type service struct {
-	db *sql.DB
+	db       *sql.DB
+	userRepo repository.UserRepository
 }
 
-var (
-	database   = os.Getenv("BLUEPRINT_DB_DATABASE")
-	password   = os.Getenv("BLUEPRINT_DB_PASSWORD")
-	username   = os.Getenv("BLUEPRINT_DB_USERNAME")
-	port       = os.Getenv("BLUEPRINT_DB_PORT")
-	host       = os.Getenv("BLUEPRINT_DB_HOST")
-	schema     = os.Getenv("BLUEPRINT_DB_SCHEMA")
-	dbInstance *service
-)
-
-func (s *service) migrate() error {
-	query := `
-	CREATE TABLE IF NOT EXISTS users (
-		id SERIAL PRIMARY KEY,
-		name TEXT NOT NULL,
-		email TEXT UNIQUE NOT NULL,
-		password TEXT NOT NULL,
-		created_at TIMESTAMPTZ DEFAULT now(),
-		updated_at TIMESTAMPTZ DEFAULT now()
-	);`
-	_, err := s.db.Exec(query)
-	return err
-}
+// func (s *service) migrate() error {
+// 	query := `
+// 	CREATE TABLE IF NOT EXISTS users (
+// 		id SERIAL PRIMARY KEY,
+// 		name TEXT NOT NULL,
+// 		email TEXT UNIQUE NOT NULL,
+// 		password TEXT NOT NULL,
+// 		created_at TIMESTAMPTZ DEFAULT now(),
+// 		updated_at TIMESTAMPTZ DEFAULT now()
+// 	);`
+// 	_, err := s.db.Exec(query)
+// 	return err
+// }
 
 func New() Service {
-	// Reuse Connection
-	if dbInstance != nil {
-		return dbInstance
+	db := NewConnection()
+	return &service{
+		db:       db,
+		userRepo: repository.NewUserRepository(db),
 	}
-	connStr := fmt.Sprintf("postgres://%s:%s@%s:%s/%s?sslmode=disable&search_path=%s", username, password, host, port, database, schema)
-	db, err := sql.Open("pgx", connStr)
-	if err != nil {
-		log.Fatal(err)
-	}
-	dbInstance = &service{
-		db: db,
-	}
-	if err := dbInstance.migrate(); err != nil {
-		log.Fatalf("failed to run migrations: %v", err)
-	}
-
-	return dbInstance
 }
 
-// Health checks the health of the database connection by pinging the database.
-// It returns a map with keys indicating various health statistics.
+func (s *service) UserRepo() repository.UserRepository {
+	return s.userRepo
+}
+
 func (s *service) Health() map[string]string {
-	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
-	defer cancel()
-
-	stats := make(map[string]string)
-
-	// Ping the database
-	err := s.db.PingContext(ctx)
-	if err != nil {
-		stats["status"] = "down"
-		stats["error"] = fmt.Sprintf("db down: %v", err)
-		log.Fatalf("db down: %v", err) // Log the error and terminate the program
-		return stats
-	}
-
-	// Database is up, add more statistics
-	stats["status"] = "up"
-	stats["message"] = "It's healthy"
-
-	// Get database stats (like open connections, in use, idle, etc.)
-	dbStats := s.db.Stats()
-	stats["open_connections"] = strconv.Itoa(dbStats.OpenConnections)
-	stats["in_use"] = strconv.Itoa(dbStats.InUse)
-	stats["idle"] = strconv.Itoa(dbStats.Idle)
-	stats["wait_count"] = strconv.FormatInt(dbStats.WaitCount, 10)
-	stats["wait_duration"] = dbStats.WaitDuration.String()
-	stats["max_idle_closed"] = strconv.FormatInt(dbStats.MaxIdleClosed, 10)
-	stats["max_lifetime_closed"] = strconv.FormatInt(dbStats.MaxLifetimeClosed, 10)
-
-	// Evaluate stats to provide a health message
-	if dbStats.OpenConnections > 40 { // Assuming 50 is the max for this example
-		stats["message"] = "The database is experiencing heavy load."
-	}
-
-	if dbStats.WaitCount > 1000 {
-		stats["message"] = "The database has a high number of wait events, indicating potential bottlenecks."
-	}
-
-	if dbStats.MaxIdleClosed > int64(dbStats.OpenConnections)/2 {
-		stats["message"] = "Many idle connections are being closed, consider revising the connection pool settings."
-	}
-
-	if dbStats.MaxLifetimeClosed > int64(dbStats.OpenConnections)/2 {
-		stats["message"] = "Many connections are being closed due to max lifetime, consider increasing max lifetime or revising the connection usage pattern."
-	}
-
-	return stats
+	return Health(s.db)
 }
 
-// CreateUser creates a new user in the database.
-func (s *service) CreateUser(ctx context.Context, user *models.User) error {
-	query := `INSERT INTO users (name, email, password, created_at, updated_at) VALUES ($1, $2, $3, NOW(), NOW()) RETURNING id, created_at, updated_at`
-	err := s.db.QueryRowContext(ctx, query, user.Name, user.Email, user.Password).Scan(&user.ID, &user.CreatedAt, &user.UpdatedAt)
-	if err != nil {
-		return fmt.Errorf("failed to create user: %w", err)
-	}
-	return nil
-}
-
-// GetUsers retrieves all users from the database.
-func (s *service) GetUsers(ctx context.Context) ([]models.User,
-	error,
-) {
-	query := `SELECT id, name, email, created_at, updated_at FROM users ORDER BY created_at DESC`
-
-	rows, err := s.db.QueryContext(ctx, query)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get users: %w", err)
-	}
-	defer rows.Close()
-
-	var users []models.User
-	for rows.Next() {
-		var user models.User
-		if err := rows.Scan(&user.ID, &user.Name, &user.Email, &user.CreatedAt, &user.UpdatedAt); err != nil {
-			return nil, fmt.Errorf("failed to scan user: %w", err)
-		}
-		users = append(users, user)
-	}
-
-	return users, nil
-}
-
-// GetUserById retrieves a user by their ID from the database.
-func (s *service) GetUserById(ctx context.Context, id int) (*models.User, error) {
-	query := `SELECT id, name, email, created_at, updated_at FROM users WHERE id = $1`
-	var user models.User
-	err := s.db.QueryRowContext(ctx, query, id).Scan(&user.ID, &user.Name, &user.Email, &user.CreatedAt, &user.UpdatedAt)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, fmt.Errorf("user not found: %w", err)
-		}
-		return nil, fmt.Errorf("failed to get user by ID: %w", err)
-	}
-	return &user, nil
-}
-
-// UpdateUser updates an existing user in the database.
-func (s *service) UpdateUser(ctx context.Context, id int, user *models.User) error {
-	query := `
-		UPDATE users
-		SET name = $1, email = $2, updated_at = NOW()
-		WHERE id = $3
-		RETURNING updated_at`
-
-	err := s.db.QueryRowContext(ctx, query, user.Name, user.Email, id).Scan(&user.UpdatedAt)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return fmt.Errorf("user with id %d not found", id)
-		}
-		return fmt.Errorf("failed to update user: %w", err)
-	}
-
-	user.ID = id
-	return nil
-}
-
-func (s *service) DeleteUser(ctx context.Context, id int) error {
-	query := `DELETE FROM users WHERE id = $1`
-	result, err := s.db.ExecContext(ctx, query, id)
-	if err != nil {
-		return fmt.Errorf("failed to delete user: %w", err)
-	}
-
-	rowsAffected, err := result.RowsAffected()
-	if err != nil {
-		return fmt.Errorf("failed to get rows affected: %w", err)
-	}
-
-	if rowsAffected == 0 {
-		return fmt.Errorf("user with id %d not found", id)
-	}
-
-	return nil
-}
-
-// Close closes the database connection.
-// It logs a message indicating the disconnection from the specific database.
-// If the connection is successfully closed, it returns nil.
-// If an error occurs while closing the connection, it returns the error.
 func (s *service) Close() error {
-	log.Printf("Disconnected from database: %s", database)
 	return s.db.Close()
 }
